Added general search
