+++ api/requirements.txt
@@ -1,6 +1,6 @@
 # Core
 pytz==2019.3
-Django==3.0.3
+Django==2.2.12
 django-configurations==2.2
 gunicorn==20.0.4
 newrelic==5.4.1.134
@@ -15,7 +15,7 @@ django-model-utils==4.0.0
 django_unique_upload==0.2.1
 
 # Rest apis
-djangorestframework==3.11.0
+djangorestframework==3.9.4
 Markdown==3.1.1
 django-filter==2.2.0
 
@@ -40,7 +40,7 @@ uwsgi==2.0.18
 
 # ElasticSearch
 django-elasticsearch-dsl-drf==0.20.8
-django-elasticsearch-dsl==7.1.1
-elasticsearch-dsl==7.1.0
-elasticsearch==7.6.0
+django-elasticsearch-dsl==6.4.2
+elasticsearch-dsl==6.4.0
+elasticsearch==6.4.0
 drf_yasg
