+++ api/resources_portal/urls.py
@@ -20,7 +20,7 @@ router.register(r"users", UserViewSet)
 router.register(r"users", UserCreateViewSet)
 router.register(r"materials", MaterialViewSet)
 
-search_router = DefaultRouter(trailing_slash=True)
+search_router = DefaultRouter(trailing_slash=False)
 search_router.register(r"materials", MaterialDocumentView, basename="")
 search_router.register(r"organizations", OrganizationDocumentView, basename="")
 search_router.register(r"users", UserDocumentView, basename="")
